+++ changelog.md
@@ -2,7 +2,9 @@
 
 ## Version wip
 - Library: perf improvement when retrieving the stories (cover not loaded when not needed)
+- Library: fix the covers that were not always removed when deleting a story
 - UI: perf improvement when displaying books (cover resized then cached)
+- UI: Sources are now editable ("Move to...")
 
 ## Version 1.4.2